refactor: remove legacy historic sql pipeline
