+++ packages/context/prompts/skills/page_triage_classifier.md
@@ -18,68 +18,6 @@ Analytics evidence (BI tools like Looker, Metabase, Tableau) is durable knowledg
 
 Treat dashboard/Look filter values, saved aggregations, calculated fields, and named tiles as candidate metric/segment definitions — they are durable. Do **not** mark BI evidence as `skip` solely because it is "configuration" or "tied to a data model"; that is exactly the durable knowledge we want to capture.
 
-Historic SQL query-history evidence is durable when usage signals show a repeated pattern worth memory work. For `signals.objectType === "historic_sql_template"`:
-
-- If `propertyHints.executions_bucket=low AND distinct_users_bucket=solo`, return `skip`. A one-off query by one user is indexed evidence, but it is too weak to produce durable knowledge candidates.
-- Else if `propertyHints.service_account_only=true AND below the frequency floor`, return `light`. Treat `executions_bucket=low` or `distinct_users_bucket=solo` as below the frequency floor for this rule. Service-account-only templates can preserve useful SQL evidence, but should not occupy a full WorkUnit unless other signals show shared human usage.
-- Otherwise apply the standard full/light/skip logic to the page excerpt. Favor `full` for shared human usage with mid or high execution volume, especially when `tables_touched`, normalized SQL, and slot classifications define a reusable metric, segment, threshold, or operational query pattern.
-
-Historic-SQL synthetic signal examples:
-
-- skip low solo template:
-
-```json
-{
-  "objectType": "historic_sql_template",
-  "propertyHints": {
-    "executions_bucket": "low",
-    "distinct_users_bucket": "solo",
-    "error_rate_bucket": "ok",
-    "recency_bucket": "active",
-    "service_account_only": "false",
-    "slot_summary": "1 constant, 1 runtime"
-  }
-}
-```
-
--> `skip`
-
-- light service-account-only template:
-
-```json
-{
-  "objectType": "historic_sql_template",
-  "propertyHints": {
-    "executions_bucket": "high",
-    "distinct_users_bucket": "solo",
-    "error_rate_bucket": "ok",
-    "recency_bucket": "active",
-    "service_account_only": "true",
-    "slot_summary": "1 constant, 0 runtime"
-  }
-}
-```
-
--> `light`
-
-- full shared human template:
-
-```json
-{
-  "objectType": "historic_sql_template",
-  "propertyHints": {
-    "executions_bucket": "high",
-    "distinct_users_bucket": "team",
-    "error_rate_bucket": "ok",
-    "recency_bucket": "active",
-    "service_account_only": "false",
-    "slot_summary": "2 constant, 1 runtime"
-  }
-}
-```
-
--> `full`
-
 Examples:
 
 - `Cold Call Script` with reusable call flow, objection handling, or positioning language -> `light` when short, `full` when multi-section or ambiguous.
