--- packages/context/src/ingest/adapters/historic-sql/__fixtures__/postgres/eviction-churn/input.json
@@ -1,146 +1,0 @@
-{
-  "name": "eviction-churn",
-  "now": "2026-05-08T12:00:00.000Z",
-  "connectionId": "warehouse",
-  "probe": {
-    "pgServerVersion": "PostgreSQL 16.4",
-    "warnings": [
-      "pg_stat_statements.max is 1000; set it to at least 5000 to reduce query-template eviction churn"
-    ]
-  },
-  "snapshot": {
-    "statsResetAt": "2026-05-08T08:00:00.000Z",
-    "deallocCount": 3,
-    "rows": [
-      {
-        "queryid": "501",
-        "userid": "11",
-        "username": "analyst",
-        "dbid": "5",
-        "database": "analytics",
-        "query": "SELECT count(*) FROM analytics.orders WHERE status = $1",
-        "calls": 20,
-        "totalExecTime": 500,
-        "meanExecTime": 25,
-        "totalRows": 40
-      }
-    ]
-  },
-  "pullConfig": {
-    "dialect": "postgres",
-    "windowDays": 90,
-    "lastSuccessfulCursor": null,
-    "serviceAccountUserPatterns": [],
-    "redactionPatterns": [],
-    "maxTemplatesPerRun": 5000,
-    "minCalls": 5
-  },
-  "analysisBySql": {
-    "SELECT count(*) FROM analytics.orders WHERE status = $1": {
-      "fingerprint": "fp_orders_status",
-      "normalizedSql": "SELECT count(*) FROM analytics.orders WHERE status = $1",
-      "tablesTouched": [
-        "analytics.orders"
-      ],
-      "literalSlots": []
-    }
-  },
-  "baseline": null,
-  "expectedBaseline": {
-    "version": 1,
-    "fetchedAt": "2026-05-08T12:00:00.000Z",
-    "statsResetAt": "2026-05-08T08:00:00.000Z",
-    "pgServerVersion": "PostgreSQL 16.4",
-    "templates": {
-      "db5_q501": {
-        "firstObservedAt": "2026-05-08T12:00:00.000Z",
-        "perUser": {
-          "11": {
-            "calls": 20,
-            "totalExecTime": 500,
-            "totalRows": 40
-          }
-        }
-      }
-    }
-  },
-  "expectedFiles": {
-    "manifest.json": {
-      "json": {
-        "source": "historic-sql",
-        "connectionId": "warehouse",
-        "dialect": "postgres",
-        "fetchedAt": "2026-05-08T12:00:00.000Z",
-        "windowStart": "2026-05-08T08:00:00.000Z",
-        "windowEnd": "2026-05-08T12:00:00.000Z",
-        "nextSuccessfulCursor": "2026-05-08T12:00:00.000Z",
-        "templateCount": 1,
-        "capped": false,
-        "warnings": [
-          "pg_stat_statements.max is 1000; set it to at least 5000 to reduce query-template eviction churn",
-          "pgss_dealloc_count:3; pg_stat_statements.max may be too low, causing template eviction churn",
-          "baseline_first_run:no_previous_pgss_baseline"
-        ],
-        "degraded": true,
-        "statsResetAt": "2026-05-08T08:00:00.000Z",
-        "baselineFirstRun": true,
-        "pgServerVersion": "PostgreSQL 16.4",
-        "deallocCount": 3,
-        "templates": [
-          {
-            "id": "db5_q501",
-            "fingerprint": "fp_orders_status",
-            "subClusterId": null,
-            "path": "templates/db5_q501/page.md"
-          }
-        ]
-      }
-    },
-    "templates/db5_q501/metadata.json": {
-      "json": {
-        "id": "db5_q501",
-        "title": "postgres · analytics.orders [db5_q501]",
-        "path": "templates/db5_q501/page.md",
-        "objectType": "historic_sql_template",
-        "lastEditedAt": null,
-        "properties": {
-          "fingerprint": "fp_orders_status",
-          "sub_cluster_id": null,
-          "dialect": "postgres",
-          "tables_touched": [
-            "analytics.orders"
-          ],
-          "literal_slots": [],
-          "triage_signals": {
-            "executions_bucket": "mid",
-            "distinct_users_bucket": "solo",
-            "error_rate_bucket": "ok",
-            "recency_bucket": "active",
-            "service_account_only": "false",
-            "runtime_bucket": "fast"
-          }
-        }
-      }
-    },
-    "templates/db5_q501/page.md": {
-      "text": "# db5_q501\n\n## Normalized SQL\n```sql\nSELECT count(*) FROM analytics.orders WHERE status = $1\n```\n\n## Tables touched\n- analytics.orders\n"
-    },
-    "templates/db5_q501/usage.json": {
-      "json": {
-        "stats": {
-          "executions": 20,
-          "distinct_users": 1,
-          "first_seen": "2026-05-08T12:00:00.000Z",
-          "last_seen": "2026-05-08T12:00:00.000Z",
-          "p50_runtime_ms": null,
-          "p95_runtime_ms": null,
-          "mean_runtime_ms": 25,
-          "error_rate": 0,
-          "rows_produced": 40
-        },
-        "literal_slots": [],
-        "samples": []
-      }
-    }
-  }
-}
