--- packages/context/src/ingest/adapters/historic-sql/__fixtures__/postgres/normal-delta/input.json
@@ -1,181 +1,0 @@
-{
-  "name": "normal-delta",
-  "now": "2026-05-08T12:00:00.000Z",
-  "connectionId": "warehouse",
-  "probe": {
-    "pgServerVersion": "PostgreSQL 16.4",
-    "warnings": []
-  },
-  "snapshot": {
-    "statsResetAt": "2026-05-08T08:00:00.000Z",
-    "deallocCount": 0,
-    "rows": [
-      {
-        "queryid": "201",
-        "userid": "11",
-        "username": "analyst",
-        "dbid": "5",
-        "database": "analytics",
-        "query": "SELECT count(*) FROM analytics.orders WHERE status = $1",
-        "calls": 12,
-        "totalExecTime": 160,
-        "meanExecTime": 13.333333333333334,
-        "totalRows": 58
-      },
-      {
-        "queryid": "201",
-        "userid": "12",
-        "username": "svc_loader",
-        "dbid": "5",
-        "database": "analytics",
-        "query": "SELECT count(*) FROM analytics.orders WHERE status = $1",
-        "calls": 5,
-        "totalExecTime": 50,
-        "meanExecTime": 10,
-        "totalRows": 25
-      }
-    ]
-  },
-  "pullConfig": {
-    "dialect": "postgres",
-    "windowDays": 90,
-    "lastSuccessfulCursor": null,
-    "serviceAccountUserPatterns": [
-      "^svc_"
-    ],
-    "redactionPatterns": [],
-    "maxTemplatesPerRun": 5000,
-    "minCalls": 5
-  },
-  "analysisBySql": {
-    "SELECT count(*) FROM analytics.orders WHERE status = $1": {
-      "fingerprint": "fp_orders_status",
-      "normalizedSql": "SELECT count(*) FROM analytics.orders WHERE status = $1",
-      "tablesTouched": [
-        "analytics.orders"
-      ],
-      "literalSlots": []
-    }
-  },
-  "baseline": {
-    "version": 1,
-    "fetchedAt": "2026-05-08T10:00:00.000Z",
-    "statsResetAt": "2026-05-08T08:00:00.000Z",
-    "pgServerVersion": "PostgreSQL 16.4",
-    "templates": {
-      "db5_q201": {
-        "firstObservedAt": "2026-05-08T09:00:00.000Z",
-        "perUser": {
-          "11": {
-            "calls": 10,
-            "totalExecTime": 100,
-            "totalRows": 50
-          },
-          "12": {
-            "calls": 5,
-            "totalExecTime": 50,
-            "totalRows": 25
-          }
-        }
-      }
-    }
-  },
-  "expectedBaseline": {
-    "version": 1,
-    "fetchedAt": "2026-05-08T12:00:00.000Z",
-    "statsResetAt": "2026-05-08T08:00:00.000Z",
-    "pgServerVersion": "PostgreSQL 16.4",
-    "templates": {
-      "db5_q201": {
-        "firstObservedAt": "2026-05-08T09:00:00.000Z",
-        "perUser": {
-          "11": {
-            "calls": 12,
-            "totalExecTime": 160,
-            "totalRows": 58
-          },
-          "12": {
-            "calls": 5,
-            "totalExecTime": 50,
-            "totalRows": 25
-          }
-        }
-      }
-    }
-  },
-  "expectedFiles": {
-    "manifest.json": {
-      "json": {
-        "source": "historic-sql",
-        "connectionId": "warehouse",
-        "dialect": "postgres",
-        "fetchedAt": "2026-05-08T12:00:00.000Z",
-        "windowStart": "2026-05-08T10:00:00.000Z",
-        "windowEnd": "2026-05-08T12:00:00.000Z",
-        "nextSuccessfulCursor": "2026-05-08T12:00:00.000Z",
-        "templateCount": 1,
-        "capped": false,
-        "warnings": [],
-        "degraded": true,
-        "statsResetAt": "2026-05-08T08:00:00.000Z",
-        "baselineFirstRun": false,
-        "pgServerVersion": "PostgreSQL 16.4",
-        "deallocCount": 0,
-        "templates": [
-          {
-            "id": "db5_q201",
-            "fingerprint": "fp_orders_status",
-            "subClusterId": null,
-            "path": "templates/db5_q201/page.md"
-          }
-        ]
-      }
-    },
-    "templates/db5_q201/metadata.json": {
-      "json": {
-        "id": "db5_q201",
-        "title": "postgres · analytics.orders [db5_q201]",
-        "path": "templates/db5_q201/page.md",
-        "objectType": "historic_sql_template",
-        "lastEditedAt": null,
-        "properties": {
-          "fingerprint": "fp_orders_status",
-          "sub_cluster_id": null,
-          "dialect": "postgres",
-          "tables_touched": [
-            "analytics.orders"
-          ],
-          "literal_slots": [],
-          "triage_signals": {
-            "executions_bucket": "low",
-            "distinct_users_bucket": "solo",
-            "error_rate_bucket": "ok",
-            "recency_bucket": "active",
-            "service_account_only": "false",
-            "runtime_bucket": "fast"
-          }
-        }
-      }
-    },
-    "templates/db5_q201/page.md": {
-      "text": "# db5_q201\n\n## Normalized SQL\n```sql\nSELECT count(*) FROM analytics.orders WHERE status = $1\n```\n\n## Tables touched\n- analytics.orders\n"
-    },
-    "templates/db5_q201/usage.json": {
-      "json": {
-        "stats": {
-          "executions": 2,
-          "distinct_users": 1,
-          "first_seen": "2026-05-08T09:00:00.000Z",
-          "last_seen": "2026-05-08T12:00:00.000Z",
-          "p50_runtime_ms": null,
-          "p95_runtime_ms": null,
-          "mean_runtime_ms": 30,
-          "error_rate": 0,
-          "rows_produced": 8
-        },
-        "literal_slots": [],
-        "samples": []
-      }
-    }
-  }
-}
