--- packages/context/src/ingest/adapters/historic-sql/__fixtures__/postgres/reset-detected/input.json
@@ -1,159 +1,0 @@
-{
-  "name": "reset-detected",
-  "now": "2026-05-08T12:00:00.000Z",
-  "connectionId": "warehouse",
-  "probe": {
-    "pgServerVersion": "PostgreSQL 16.4",
-    "warnings": []
-  },
-  "snapshot": {
-    "statsResetAt": "2026-05-08T11:00:00.000Z",
-    "deallocCount": 0,
-    "rows": [
-      {
-        "queryid": "301",
-        "userid": "11",
-        "username": "analyst",
-        "dbid": "5",
-        "database": "analytics",
-        "query": "SELECT count(*) FROM analytics.orders WHERE status = $1",
-        "calls": 3,
-        "totalExecTime": 90,
-        "meanExecTime": 30,
-        "totalRows": 9
-      }
-    ]
-  },
-  "pullConfig": {
-    "dialect": "postgres",
-    "windowDays": 90,
-    "lastSuccessfulCursor": null,
-    "serviceAccountUserPatterns": [],
-    "redactionPatterns": [],
-    "maxTemplatesPerRun": 5000,
-    "minCalls": 5
-  },
-  "analysisBySql": {
-    "SELECT count(*) FROM analytics.orders WHERE status = $1": {
-      "fingerprint": "fp_orders_status",
-      "normalizedSql": "SELECT count(*) FROM analytics.orders WHERE status = $1",
-      "tablesTouched": [
-        "analytics.orders"
-      ],
-      "literalSlots": []
-    }
-  },
-  "baseline": {
-    "version": 1,
-    "fetchedAt": "2026-05-08T10:00:00.000Z",
-    "statsResetAt": "2026-05-08T08:00:00.000Z",
-    "pgServerVersion": "PostgreSQL 16.4",
-    "templates": {
-      "db5_q301": {
-        "firstObservedAt": "2026-05-08T09:00:00.000Z",
-        "perUser": {
-          "11": {
-            "calls": 100,
-            "totalExecTime": 1000,
-            "totalRows": 500
-          }
-        }
-      }
-    }
-  },
-  "expectedBaseline": {
-    "version": 1,
-    "fetchedAt": "2026-05-08T12:00:00.000Z",
-    "statsResetAt": "2026-05-08T11:00:00.000Z",
-    "pgServerVersion": "PostgreSQL 16.4",
-    "templates": {
-      "db5_q301": {
-        "firstObservedAt": "2026-05-08T12:00:00.000Z",
-        "perUser": {
-          "11": {
-            "calls": 3,
-            "totalExecTime": 90,
-            "totalRows": 9
-          }
-        }
-      }
-    }
-  },
-  "expectedFiles": {
-    "manifest.json": {
-      "json": {
-        "source": "historic-sql",
-        "connectionId": "warehouse",
-        "dialect": "postgres",
-        "fetchedAt": "2026-05-08T12:00:00.000Z",
-        "windowStart": "2026-05-08T10:00:00.000Z",
-        "windowEnd": "2026-05-08T12:00:00.000Z",
-        "nextSuccessfulCursor": "2026-05-08T12:00:00.000Z",
-        "templateCount": 1,
-        "capped": false,
-        "warnings": [
-          "baseline_reset:stats_reset advanced from 2026-05-08T08:00:00.000Z to 2026-05-08T11:00:00.000Z"
-        ],
-        "degraded": true,
-        "statsResetAt": "2026-05-08T11:00:00.000Z",
-        "baselineFirstRun": true,
-        "pgServerVersion": "PostgreSQL 16.4",
-        "deallocCount": 0,
-        "templates": [
-          {
-            "id": "db5_q301",
-            "fingerprint": "fp_orders_status",
-            "subClusterId": null,
-            "path": "templates/db5_q301/page.md"
-          }
-        ]
-      }
-    },
-    "templates/db5_q301/metadata.json": {
-      "json": {
-        "id": "db5_q301",
-        "title": "postgres · analytics.orders [db5_q301]",
-        "path": "templates/db5_q301/page.md",
-        "objectType": "historic_sql_template",
-        "lastEditedAt": null,
-        "properties": {
-          "fingerprint": "fp_orders_status",
-          "sub_cluster_id": null,
-          "dialect": "postgres",
-          "tables_touched": [
-            "analytics.orders"
-          ],
-          "literal_slots": [],
-          "triage_signals": {
-            "executions_bucket": "mid",
-            "distinct_users_bucket": "solo",
-            "error_rate_bucket": "ok",
-            "recency_bucket": "active",
-            "service_account_only": "false",
-            "runtime_bucket": "fast"
-          }
-        }
-      }
-    },
-    "templates/db5_q301/page.md": {
-      "text": "# db5_q301\n\n## Normalized SQL\n```sql\nSELECT count(*) FROM analytics.orders WHERE status = $1\n```\n\n## Tables touched\n- analytics.orders\n"
-    },
-    "templates/db5_q301/usage.json": {
-      "json": {
-        "stats": {
-          "executions": 3,
-          "distinct_users": 1,
-          "first_seen": "2026-05-08T12:00:00.000Z",
-          "last_seen": "2026-05-08T12:00:00.000Z",
-          "p50_runtime_ms": null,
-          "p95_runtime_ms": null,
-          "mean_runtime_ms": 30,
-          "error_rate": 0,
-          "rows_produced": 9
-        },
-        "literal_slots": [],
-        "samples": []
-      }
-    }
-  }
-}
