--- packages/context/src/ingest/adapters/historic-sql/__fixtures__/postgres/version-change/input.json
@@ -1,159 +1,0 @@
-{
-  "name": "version-change",
-  "now": "2026-05-08T12:00:00.000Z",
-  "connectionId": "warehouse",
-  "probe": {
-    "pgServerVersion": "PostgreSQL 16.4",
-    "warnings": []
-  },
-  "snapshot": {
-    "statsResetAt": "2026-05-08T08:00:00.000Z",
-    "deallocCount": 0,
-    "rows": [
-      {
-        "queryid": "401",
-        "userid": "11",
-        "username": "analyst",
-        "dbid": "5",
-        "database": "analytics",
-        "query": "SELECT count(*) FROM analytics.orders WHERE status = $1",
-        "calls": 4,
-        "totalExecTime": 80,
-        "meanExecTime": 20,
-        "totalRows": 8
-      }
-    ]
-  },
-  "pullConfig": {
-    "dialect": "postgres",
-    "windowDays": 90,
-    "lastSuccessfulCursor": null,
-    "serviceAccountUserPatterns": [],
-    "redactionPatterns": [],
-    "maxTemplatesPerRun": 5000,
-    "minCalls": 5
-  },
-  "analysisBySql": {
-    "SELECT count(*) FROM analytics.orders WHERE status = $1": {
-      "fingerprint": "fp_orders_status",
-      "normalizedSql": "SELECT count(*) FROM analytics.orders WHERE status = $1",
-      "tablesTouched": [
-        "analytics.orders"
-      ],
-      "literalSlots": []
-    }
-  },
-  "baseline": {
-    "version": 1,
-    "fetchedAt": "2026-05-08T10:00:00.000Z",
-    "statsResetAt": "2026-05-08T08:00:00.000Z",
-    "pgServerVersion": "PostgreSQL 15.7",
-    "templates": {
-      "db5_q401": {
-        "firstObservedAt": "2026-05-08T09:00:00.000Z",
-        "perUser": {
-          "11": {
-            "calls": 100,
-            "totalExecTime": 1000,
-            "totalRows": 500
-          }
-        }
-      }
-    }
-  },
-  "expectedBaseline": {
-    "version": 1,
-    "fetchedAt": "2026-05-08T12:00:00.000Z",
-    "statsResetAt": "2026-05-08T08:00:00.000Z",
-    "pgServerVersion": "PostgreSQL 16.4",
-    "templates": {
-      "db5_q401": {
-        "firstObservedAt": "2026-05-08T12:00:00.000Z",
-        "perUser": {
-          "11": {
-            "calls": 4,
-            "totalExecTime": 80,
-            "totalRows": 8
-          }
-        }
-      }
-    }
-  },
-  "expectedFiles": {
-    "manifest.json": {
-      "json": {
-        "source": "historic-sql",
-        "connectionId": "warehouse",
-        "dialect": "postgres",
-        "fetchedAt": "2026-05-08T12:00:00.000Z",
-        "windowStart": "2026-05-08T10:00:00.000Z",
-        "windowEnd": "2026-05-08T12:00:00.000Z",
-        "nextSuccessfulCursor": "2026-05-08T12:00:00.000Z",
-        "templateCount": 1,
-        "capped": false,
-        "warnings": [
-          "baseline_reset:pg_server_major changed from 15 to 16"
-        ],
-        "degraded": true,
-        "statsResetAt": "2026-05-08T08:00:00.000Z",
-        "baselineFirstRun": true,
-        "pgServerVersion": "PostgreSQL 16.4",
-        "deallocCount": 0,
-        "templates": [
-          {
-            "id": "db5_q401",
-            "fingerprint": "fp_orders_status",
-            "subClusterId": null,
-            "path": "templates/db5_q401/page.md"
-          }
-        ]
-      }
-    },
-    "templates/db5_q401/metadata.json": {
-      "json": {
-        "id": "db5_q401",
-        "title": "postgres · analytics.orders [db5_q401]",
-        "path": "templates/db5_q401/page.md",
-        "objectType": "historic_sql_template",
-        "lastEditedAt": null,
-        "properties": {
-          "fingerprint": "fp_orders_status",
-          "sub_cluster_id": null,
-          "dialect": "postgres",
-          "tables_touched": [
-            "analytics.orders"
-          ],
-          "literal_slots": [],
-          "triage_signals": {
-            "executions_bucket": "mid",
-            "distinct_users_bucket": "solo",
-            "error_rate_bucket": "ok",
-            "recency_bucket": "active",
-            "service_account_only": "false",
-            "runtime_bucket": "fast"
-          }
-        }
-      }
-    },
-    "templates/db5_q401/page.md": {
-      "text": "# db5_q401\n\n## Normalized SQL\n```sql\nSELECT count(*) FROM analytics.orders WHERE status = $1\n```\n\n## Tables touched\n- analytics.orders\n"
-    },
-    "templates/db5_q401/usage.json": {
-      "json": {
-        "stats": {
-          "executions": 4,
-          "distinct_users": 1,
-          "first_seen": "2026-05-08T12:00:00.000Z",
-          "last_seen": "2026-05-08T12:00:00.000Z",
-          "p50_runtime_ms": null,
-          "p95_runtime_ms": null,
-          "mean_runtime_ms": 20,
-          "error_rate": 0,
-          "rows_produced": 8
-        },
-        "literal_slots": [],
-        "samples": []
-      }
-    }
-  }
-}
